# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: number;
  first_name: string;
  second_name: string;
  third_name: string;
  full_name_id: string;
  phone: string;
  skills: string;
  /** @nullable */
  profile_photo_path: string | null;
  /** @nullable */
  id_front_path: string | null;
  /** @nullable */
  id_back_path: string | null;
  created_at: string;
}

export interface CreateEmployeeBody {
  first_name: string;
  second_name: string;
  third_name: string;
  full_name_id: string;
  phone: string;
  skills: string;
  profile_photo_path: string;
  id_front_path: string;
  id_back_path: string;
}

export interface LoginBody {
  phone: string;
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}