# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Portal API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all employees (admin only)
 */
export const ListEmployeesResponseItem = zod.object({
  id: zod.number(),
  first_name: zod.string(),
  second_name: zod.string(),
  third_name: zod.string(),
  full_name_id: zod.string(),
  phone: zod.string(),
  skills: zod.string(),
  profile_photo_path: zod.string().nullable(),
  id_front_path: zod.string().nullable(),
  id_back_path: zod.string().nullable(),
  created_at: zod.coerce.date(),
});
export const ListEmployeesResponse = zod.array(ListEmployeesResponseItem);

/**
 * @summary Submit employee registration
 */
export const CreateEmployeeBody = zod.object({
  first_name: zod.string(),
  second_name: zod.string(),
  third_name: zod.string(),
  full_name_id: zod.string(),
  phone: zod.string(),
  skills: zod.string(),
  profile_photo_path: zod.string(),
  id_front_path: zod.string(),
  id_back_path: zod.string(),
});

/**
 * @summary Get a single employee by ID (admin only)
 */
export const GetEmployeeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetEmployeeResponse = zod.object({
  id: zod.number(),
  first_name: zod.string(),
  second_name: zod.string(),
  third_name: zod.string(),
  full_name_id: zod.string(),
  phone: zod.string(),
  skills: zod.string(),
  profile_photo_path: zod.string().nullable(),
  id_front_path: zod.string().nullable(),
  id_back_path: zod.string().nullable(),
  created_at: zod.coerce.date(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  phone: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary Check current session
 */
export const GetAuthSessionResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});

/**
 * @summary Serve a stored object
 */
export const GetStorageObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});